+++ lib/application/usecases/get_group_members_by_group_id_usecase.dart
@@ -1,0 +1,12 @@
+import '../../domain/entities/group_member.dart';
+import '../../domain/repositories/group_member_repository.dart';
+
+class GetGroupMembersByGroupIdUsecase {
+  final GroupMemberRepository _groupMemberRepository;
+
+  GetGroupMembersByGroupIdUsecase(this._groupMemberRepository);
+
+  Future<List<GroupMember>> execute(String groupId) async {
+    return await _groupMemberRepository.getGroupMembersByGroupId(groupId);
+  }
+}
